requirements.txt
Django==4.2.25
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
stripe==8.0.0
python-decouple==3.8
pytz==2023.3
